improve layout of track settings component
